cli/requirements.txt
requests>=2.25.0
pyyaml>=5.4.0
click>=8.0.0
rich>=10.0.0
tabulate>=0.8.9
pytest>=6.0.0
pytest-cov>=2.10.0 